requirements.txt
netcdf4
h5netcdf
#numpy~=1.20.1
numpy
matplotlib~=3.3.4
#scipy~=1.6.1
scipy
#xarray~=0.18.0
xarray
#datacube~=1.8.5
sklearn~=0.0
scikit-learn~=0.24.2
#GDAL~=3.3.1
rasterio~=1.2.6
#pandas~=1.2.2
pandas
seaborn~=0.11.1
#scikit-image~=0.18.2
scikit-image
Fiona~=1.8.19
Shapely~=1.7.1
pyproj~=3.0.1
bokeh~=2.3.3
click~=7.1.2
PyYAML~=5.4.1
geopandas~=0.9.0
dask~=2021.3.0
#torch~=1.9.0
pycm~=3.3
pathlib
tslearn
imgkit